fix(identity): require explicit config for did:web overwrites
