+++ agent_identity/src/document/event.rs
@@ -1,6 +1,7 @@
 use super::aggregate::{IotaMetadata, Status};
 use agent_shared::config::SupportedDidMethod;
 use cqrs_es::DomainEvent;
+use identity_did::CoreDID;
 use identity_document::document::CoreDocument;
 use jsonwebtoken::Algorithm;
 use serde::{Deserialize, Serialize};
@@ -16,8 +17,9 @@ pub enum DocumentEvent {
         with_fixed_algorithm: Option<Algorithm>,
         iota_metadata: Option<IotaMetadata>,
     },
-    DocumentIdentityChanged {
+    DocumentDidWebOverwritten {
         document_id: String,
+        previous_did: CoreDID,
         document: CoreDocument,
     },
     PublicKeyUpdated {
